% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_check.R
\name{convergence_check}
\alias{convergence_check}
\title{Print convergence diagnostics}
\usage{
convergence_check(mod)
}
\arguments{
\item{mod}{A BBT model}
}
\value{
The output from cmdstan_diagnose() from the cmdstanr package
}
\description{
Print convergence diagnostics
}
\examples{
\donttest{
ss <- ll[1:80, 1:6] # first 20 data sets and 5 algorithms

m1 <- bbtcomp(ss)
convergence_check(m1)
}
}
