% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmcbbt.R
\name{mcmcbbt}
\alias{mcmcbbt}
\title{Fit a BBT model by running MCMC on a wintable}
\usage{
mcmcbbt(tablex, hyper_prior = 0, scale = 0.5, use_davidson = FALSE, ...)
}
\arguments{
\item{tablex}{A wintable}

\item{hyper_prior}{The code of the hyper-prior for the sigma

\if{html}{\out{<div class="sourceCode">}}\preformatted{* 0 = lognormal(0.5) - the default,
* 1 = lognormal(scale),
* 2 = cauchy(scale),
* 3 = normal(scale)
}\if{html}{\out{</div>}}}

\item{scale}{The scale of the hyper prior for the sigma parameter}

\item{use_davidson}{Whether to use the Davidson model to deal with ties}

\item{...}{Further parameters to the MCMC (cmdstanr model and fit)}
}
\value{
A fitted BBT model
}
\description{
Fit a BBT model by running MCMC on a wintable
}
\section{CmdStan setup}{

Installing bbtcomp installs the \code{cmdstanr} R package, but it
cannot install or verify CmdStan itself (the compiled Stan backend).
If CmdStan cannot be found, this function stops with setup
instructions (see \code{\link{check_setup}}) instead of a deep
cmdstanr error.
}

\examples{
\donttest{
ss <- ll[1:80, 1:6] # first 20 data sets and 5 algorithms

wintable <- make_wintable(ss, lrope = FALSE)

options(mc.cores = parallel::detectCores(logical = FALSE))

m1 <- mcmcbbt(wintable, hyper_prior = 0, scale = 0.5,
     use_davidson = TRUE)

m2 <- mcmcbbt(wintable, iter_sampling = 2000, seed = 1234)
}
}
