% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pwin.R
\name{plot_pwin}
\alias{plot_pwin}
\title{Plot the probabilities of an algorithm being better than another}
\usage{
plot_pwin(
  modout,
  selected = NULL,
  control = NULL,
  rope = c(0.45, 0.55),
  hdi = 0.89
)
}
\arguments{
\item{modout}{A BBT model}

\item{selected}{List of algorithm names to be printed. NULL for all}

\item{control}{An algorithm name, to be seen as the control. Only the
probabilities of the other algorithms against the control will be
plotted}

\item{rope}{The low and the high values of ROPE}

\item{hdi}{The HDI probability to indicate in the plot.}
}
\value{
A ggplot2 plot
}
\description{
Plot the probabilities of an algorithm being better than another
}
\examples{
\donttest{
m1 = bbtcomp(ll, lrope = FALSE, deal_with_ties = "davidson")
plot_pwin(m1, control = "rf")
plot_pwin(m1, selected = c("svm","lda","gbm","passive"), rope = c(0.48, 0.52))
plot_pwin(m1,  hdi = 0.95)
}
}
